Revert moving toDto from interventionCatalogue class
