+++ src/main/kotlin/uk/gov/justice/digital/hmpps/findandreferanintervention/controller/InterventionController.kt
@@ -4,7 +4,6 @@ import com.microsoft.applicationinsights.TelemetryClient
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.web.PageableDefault
-import org.springframework.data.web.config.EnableSpringDataWebSupport
 import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.security.access.prepost.PreAuthorize
@@ -23,7 +22,6 @@ import uk.gov.justice.digital.hmpps.findandreferanintervention.service.Intervent
 import java.util.UUID
 
 @RestController
-@EnableSpringDataWebSupport(pageSerializationMode = EnableSpringDataWebSupport.PageSerializationMode.VIA_DTO)
 @PreAuthorize("hasRole('ROLE_FIND_AND_REFER_AN_INTERVENTION_API__FAR_UI__WR')")
 class InterventionController(
   private val interventionService: InterventionService,
