+++ src/main/kotlin/uk/gov/justice/digital/hmpps/findandreferanintervention/dto/InterventionCatalogueDto.kt
@@ -1,7 +1,6 @@
 package uk.gov.justice.digital.hmpps.findandreferanintervention.dto
 
 import com.fasterxml.jackson.annotation.JsonInclude
-import uk.gov.justice.digital.hmpps.findandreferanintervention.jpa.entity.InterventionCatalogue
 import uk.gov.justice.digital.hmpps.findandreferanintervention.jpa.entity.InterventionType
 import uk.gov.justice.digital.hmpps.findandreferanintervention.jpa.entity.SettingType
 import java.util.UUID
@@ -24,34 +23,3 @@ data class InterventionCatalogueDto(
   val suitableForPeopleWithLearningDifficulties: String?,
   val equivalentNonLdcProgramme: String?,
 )
-
-fun InterventionCatalogue.toDto(): InterventionCatalogueDto {
-  val deliveryMethodDtos =
-    this.deliveryMethods.map { DeliveryMethodDto.fromEntity(it) }
-  val deliveryMethodSettingList =
-    deliveryMethodDtos.flatMap { methodDto ->
-      methodDto.deliveryMethodSettings.map { settingDto -> settingDto.setting }
-    }
-  return InterventionCatalogueDto(
-    id = this.id,
-    criminogenicNeeds =
-    this.criminogenicNeeds.map {
-      CriminogenicNeedDto.fromEntity(it).need
-    },
-    title = this.name,
-    description = this.shortDescription,
-    interventionType = this.interventionType,
-    setting = deliveryMethodSettingList,
-    allowsMales = this.personalEligibility?.males!!,
-    allowsFemales = this.personalEligibility?.females!!,
-    riskCriteria =
-    this.riskConsideration?.let {
-      RiskConsiderationDto.fromEntity(it).listOfRisks()
-    },
-    attendanceType = deliveryMethodDtos.mapNotNull { methodDto -> methodDto.attendanceType },
-    deliveryFormat = deliveryMethodDtos.mapNotNull { methodDto -> methodDto.deliveryFormat },
-    timeToComplete = this.timeToComplete,
-    suitableForPeopleWithLearningDifficulties = this.specialEducationalNeeds?.learningDisabilityCateredFor,
-    equivalentNonLdcProgramme = this.specialEducationalNeeds?.equivalentNonLdcProgrammeGuide,
-  )
-}
