+++ src/main/kotlin/uk/gov/justice/digital/hmpps/findandreferanintervention/dto/RiskConsiderationDto.kt
@@ -35,7 +35,7 @@ data class RiskConsiderationDto(
         pnaGuide,
         rsrGuide,
         roshLevel?.toString(),
-      )
+      ).sorted()
     }
 
   companion object {
